# server/src/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"TransactionResponse": {
				"properties": {
					"account_holder": {
						"type": "string"
					},
					"iban": {
						"type": "string"
					},
					"date": {
						"type": "string"
					},
					"amount": {
						"type": "string"
					},
					"note": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"account_holder",
					"iban",
					"date",
					"amount",
					"note",
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Transaction": {
				"properties": {
					"account_holder": {
						"type": "string"
					},
					"iban": {
						"type": "string"
					},
					"date": {
						"type": "string"
					},
					"amount": {
						"type": "string"
					},
					"note": {
						"type": "string"
					}
				},
				"required": [
					"account_holder",
					"iban",
					"date",
					"amount",
					"note"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "transfer-app-server",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/transaction/getAll": {
			"get": {
				"operationId": "GetAllTransactions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TransactionResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Transaction"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/transaction": {
			"post": {
				"operationId": "CreateTransaction",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TransactionResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Transaction"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Transaction"
							}
						}
					}
				}
			}
		},
		"/api/transaction/{id}": {
			"put": {
				"operationId": "UpdateTransaction",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TransactionResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Transaction"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Transaction"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteTransaction",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Transaction"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}